[SNAP-1478](snap-product-library-ui-v2)(snap-worldwind): Integrate the SNAP World Wind module in SNAP Product Library to remove the dependencies' redundancy (the NASA World Wind library and its native libraries).
