+++ snap-worldwind/pom.xml
@@ -65,6 +65,11 @@
             <groupId>org.esa.snap</groupId>
             <artifactId>snap-engine-utilities</artifactId>
         </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>snap-remote-products-repository</artifactId>
+            <version>${project.version}</version>
+        </dependency>
 
         <dependency>
             <groupId>gov.nasa</groupId>
