+++ snap-worldwind/src/main/java/org/esa/snap/worldwind/productlibrary/FixingPlaceNameLayer.java
@@ -16,7 +16,7 @@
  *
  */
 
-package org.esa.snap.product.library.ui.v2.worldwind;
+package org.esa.snap.worldwind.productlibrary;
 
 import gov.nasa.worldwind.layers.Earth.NASAWFSPlaceNameLayer;
 import gov.nasa.worldwind.render.DeclutteringTextRenderer;
