+++ snap-worldwind/src/main/java/org/esa/snap/worldwind/productlibrary/Map2DMouseHandler.java
@@ -1,4 +1,4 @@
-package org.esa.snap.product.library.ui.v2.worldwind;
+package org.esa.snap.worldwind.productlibrary;
 
 import com.bc.ceres.glayer.swing.LayerCanvas;
 import org.apache.commons.math3.util.FastMath;
