+++ snap-worldwind/src/main/java/org/esa/snap/worldwind/productlibrary/Rectangle3DSelection.java
@@ -1,4 +1,4 @@
-package org.esa.snap.product.library.ui.v2.worldwind;
+package org.esa.snap.worldwind.productlibrary;
 
 import gov.nasa.worldwind.WorldWindow;
 import gov.nasa.worldwind.geom.Sector;
