+++ snap-worldwind/src/main/java/org/esa/snap/worldwind/productlibrary/WorldMap2DPanel.java
@@ -1,4 +1,4 @@
-package org.esa.snap.product.library.ui.v2.worldwind;
+package org.esa.snap.worldwind.productlibrary;
 
 import com.bc.ceres.binding.PropertyContainer;
 import com.bc.ceres.glayer.Layer;
